🚚  -> (base)/my-shop/(home) -> base/(home)로 변경
sub-issues:
- [NAILCASE-394](https://nailcase.atlassian.net/browse/NAILCASE-394)
